requirements.txt
altgraph==0.17.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
future==0.18.2
idna==3.3
Naked==0.1.31
pefile==2021.9.3
pycparser==2.21
pycryptodome==3.14.1
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.2
pywin32==303
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.27.1
shellescape==3.8.1
urllib3==1.26.8
